removing bundles, was a bad idea, just insert rig and actions to a camera
