+++ examples/follow.rs
@@ -25,13 +25,16 @@ fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
         .insert(Sheep)
         .id();
 
-    commands.spawn_bundle(DollyCameraBundle {
-        rig: Rig::default()
-            .with(Anchor::new(sheep))
-            .with(Arm::new(Vec3::new(0.0, 2.5, -3.5)))
-            .with(LookAt::new(sheep, Vec3::new(0.0, 1.0, 1.0))),
-        ..Default::default()
-    });
+    commands
+        .spawn_bundle(PerspectiveCameraBundle {
+            ..Default::default()
+        })
+        .insert(
+            Rig::default()
+                .with(Anchor::new(sheep))
+                .with(Arm::new(Vec3::new(0.0, 2.5, -3.5)))
+                .with(LookAt::new(sheep, Vec3::new(0.0, 1.0, 1.0))),
+        );
 
     info!(" Use Z and X to turn the sheep");
 }
