+++ src/lib.rs
@@ -1,13 +1,12 @@
 #![feature(derive_default_enum)]
-pub mod bundle;
-pub mod control;
+pub mod actions;
 pub mod rig;
 use bevy::prelude::*;
-use control::*;
-use rig::*;
+pub use actions::*;
+pub use rig::*;
 
 pub mod prelude {
-    pub use crate::{bundle::*, control::*, rig::*, *};
+    pub use crate::{actions::*, rig::*, *};
 }
 
 pub struct DollyPlugin;
@@ -21,7 +20,7 @@ impl Plugin for DollyPlugin {
 
             // These are only use for camera control system
             .init_resource::<DollyControlConfig>()
-            .add_system_to_stage(CoreStage::PreUpdate, update_control_system);
+            .add_system_to_stage(CoreStage::PreUpdate, actions::update_control_system);
     }
 }
 
@@ -31,8 +30,6 @@ fn init_rig_system(mut query: Query<(&mut Transform, &mut Rig), Added<Rig>>) {
     for (transform, mut rig) in query.iter_mut() {
         // set init target for the transform
         rig.target = *transform;
-
-        
     }
 }
 
@@ -62,7 +59,5 @@ fn apply_rigs_system(time: Res<Time>, mut query: Query<(&mut Transform, &mut Rig
         let delta_seconds = time.delta_seconds();
         // Apply Rigs
         *transform = rig.update(&transform, delta_seconds);
-
-
     }
 }
